+++ tests/unit/generators/conf_files/test_create_alert_actions_conf.py
@@ -1,21 +1,28 @@
 import shutil
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch
 from splunk_add_on_ucc_framework.generators.conf_files import AlertActionsConf
 from textwrap import dedent
 
 
-def test_set_attributes_global_config_with_empty_alerts(
-    global_config_for_alerts,
-    input_dir,
-    output_dir,
+@patch.object(shutil, "copy")
+def test_custom_icon_file_name(
+    mock_copy, global_config_for_alerts, input_dir, output_dir
 ):
-    global_config_for_alerts = MagicMock()
-    global_config_for_alerts.alerts = []
-
     alert_action_conf = AlertActionsConf(
         global_config_for_alerts, input_dir, output_dir
     )
-    alert_action_conf._set_attributes()
+
+    assert "icon_path = dev_icon.png" in alert_action_conf.alerts["test_alert_default"]
+
+
+def test_init_global_config_with_empty_alerts(
+    global_config_only_configuration,
+    input_dir,
+    output_dir,
+):
+    alert_action_conf = AlertActionsConf(
+        global_config_only_configuration, input_dir, output_dir
+    )
 
     assert alert_action_conf.alerts == {}
     assert alert_action_conf.alerts_spec == {}
@@ -59,7 +66,7 @@ def test_generate_conf(mock_copy, global_config_for_alerts, input_dir, output_di
                 "is_custom = 1",
                 "payload_format = json",
                 "[test_alert_default]",
-                "icon_path = alerticon.png",
+                "icon_path = dev_icon.png",
                 "label = Test Alert Default",
                 "description = Description for test Alert Action",
                 (