+++ tests/unit/generators/python_files/test_create_custom_command_python.py
@@ -25,6 +25,10 @@ def custom_search_commands():
                     "validate": {"type": "Integer", "minimum": 5, "maximum": 10},
                 },
                 {
+                    "name": "max_word",
+                    "validate": {"type": "Integer", "maximum": 100},
+                },
+                {
                     "name": "age",
                     "validate": {"type": "Integer", "minimum": 18},
                 },
@@ -36,7 +40,7 @@ def custom_search_commands():
     ]
 
 
-def test_set_attributes_without_custom_command(
+def test_init_without_custom_command(
     global_config_only_configuration,
     input_dir,
     output_dir,
@@ -49,7 +53,7 @@ def test_set_attributes_without_custom_command(
     assert custom_command.commands_info == []
 
 
-def test_set_attributes(
+def test_init(
     global_config_all_json,
     input_dir,
     output_dir,
@@ -72,6 +76,7 @@ def test_set_attributes(
             "list_arg": [
                 "count = Option(name='count', require=True, "
                 "validate=validators.Integer(minimum=5, maximum=10))",
+                "max_word = Option(name='max_word', require=False, validate=validators.Integer(maximum=100))",
                 "age = Option(name='age', require=False, validate=validators.Integer(minimum=18))",
                 "text = Option(name='text', require=True, default='test_text')",
                 "contains = Option(name='contains', require=False)",